server/requirements/base.txt
Django==5.2.1
wagtail==6.4.1
psycopg2-binary==2.9.10
python-decouple==3.8

djangorestframework==3.15.2
dj-rest-auth==7.0.1
djangorestframework-simplejwt==5.5.0
django-allauth==65.8.0
django-cors-headers==4.7.0
django-extensions==4.1
spotipy==2.24.0
pycountry==24.6.1
google-generativeai==0.8.5

gunicorn==23.0.0

